Move `get_source_inputs` (#10415)
