+++ keras/utils/__init__.py
@@ -18,6 +18,7 @@ from .generic_utils import serialize_keras_object
 from .generic_utils import deserialize_keras_object
 from .generic_utils import Progbar
 from .layer_utils import convert_all_kernels_in_model
+from .layer_utils import get_source_inputs
 from .layer_utils import print_summary
 from .vis_utils import plot_model
 from .np_utils import to_categorical
